binutils: bump version to 2.44 for fix RPi vulkan retroarch service failed (#2125)
* binutils: update to 2.42

* binutils: update to 2.43

ann:
- https://lists.gnu.org/archive/html/info-gnu/2024-08/msg00001.html

* binutils: update to 2.43.1

ann:
- https://lists.gnu.org/archive/html/info-gnu/2024-08/msg00003.html

* binutils: update to 2.44

release notes:
- https://sourceware.org/pipermail/binutils/2025-February/139195.html

In a change to our previous practice, in this release the
binutils-2.44.tar tarball does not contain the sources for the gold
linker.  This is because the gold linker is now deprecated and will
eventually be removed unless volunteers step forward and offer to
continue development and maintenance.

The gold sources can be found in the binutils-with-gold-2.44.tar
tarballs.  Going forward, odd numbered releases of the binutils (2.45,
2.47, etc) will just have the binutils.2.xx.tar tarballs, whilst even
numbered releases will have both the binutils-2.xx.tar and the
binutils-with-gold-2.xx.tar tarballs.  Eventually this will stop and
gold will be dropped altogether.

---------

Co-authored-by: Rudi Heitbaum <rudi@heitbaum.com>